Funktionsauslagerung / Refactoring / Bugfixes / Code Convention
