+++ ProjektKlausurgenerator/Muster Fragen CSV-Datei.csv
@@ -1,0 +1,3 @@
+Freitext;Hier kommt der Fragentext;schwer;20;Java Aufbau;Stichwort1 Stichwort2 Stichwort3
+
+Multichoice;Hier kommt die Frage;leicht;10;Java Aufbau;Antwort1;True;Antwort2;False;Antwort3;False;Antwort4;False
