--- ProjektKlausurgenerator/fragen-Sicherung.csv
@@ -1,3 +1,0 @@
-Freitext;Frage1;schwer;20;Java Aufbau
-Multichoice;Frage2;leicht;20;Grundlagen;Antwort1;Antwort2;Antwort3;Antwort4
-Multichoice;Frage3;mittel;10;Java Aufbau;Antwort1;Antwort2;Antwort3;Antwort4
