+++ ProjektKlausurgenerator/src/de/arbeitsagentur/ProjektKlausurgenerator/Oberflaeche/HilfeFenster.java
@@ -3,27 +3,39 @@ package de.arbeitsagentur.ProjektKlausurgenerator.Oberflaeche;
 import java.awt.Color;
 import java.awt.Cursor;
 import java.awt.Font;
+
 import javax.swing.ImageIcon;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTextArea;
 
+/**
+ * Hilfefenster der Anwendung mit Hilfetexten zur groben Erlauterung.
+ * 
+ * @author Nico & Anna & Daniel
+ *
+ */
 public class HilfeFenster {
+	
+	// *** Eigenschaften *** //
 
 	private JFrame frame;
-	private JPanel panel;
-	private final String hilfetext = "\nPrüfungstool der Gruppe F18."
+	private static final String HILFETEXT = "\nPrüfungstool der Gruppe F18."
 			+ "\n\nKlausur Importieren:\nHier kann eine existierende Klausur als CSV eingelesen werden."
 			+ "\n\nKlausur Exportieren:\nHier kann eine erstellte Klausur als CSV oder als PDF exportiert werden."
 			+ "\n\nKlausur Erstellen:\nEs öffnet sich ein neues Fenster mit einer Tabellenansicht, in der man eine Klausur aus bereits existierende Fragen zusammenstellen kann."
 			+ "\n\nFrage Hinzufügen:\nIn diesem Fenster erstellt man neu Fragen.";
 
+	// *** Konstruktor *** //
+	
 	public HilfeFenster() {
 		initialize();
 
 		frame.setVisible(true);
 	}
+	
+	// *** Methoden *** //
 
 	private void initialize() {
 		frame = new JFrame();
@@ -34,7 +46,7 @@ public class HilfeFenster {
 		frame.setTitle("Hilfe Klausurtool");
 		frame.getContentPane().setLayout(null);
 
-		panel = new JPanel();
+		JPanel panel = new JPanel();
 		panel.setBackground(Color.WHITE);
 		panel.setBounds(0, 0, 429, 601);
 		frame.getContentPane().add(panel);
@@ -50,7 +62,7 @@ public class HilfeFenster {
 		JTextArea textAreaHilfetext = new JTextArea();
 		textAreaHilfetext.setFont(new Font("Arial", Font.PLAIN, 15));
 		textAreaHilfetext.setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
-		textAreaHilfetext.setText(hilfetext);
+		textAreaHilfetext.setText(HILFETEXT);
 		textAreaHilfetext.setEditable(false);
 		textAreaHilfetext.setLineWrap(true);
 		textAreaHilfetext.setWrapStyleWord(true);
@@ -58,7 +70,7 @@ public class HilfeFenster {
 		panel.add(textAreaHilfetext);
 
 		// Label-Bild
-		ImageIcon imageIcon = GuiUtils.getScaledImageIcon("images/fragezeichen.PNG", 39, 39);
+		ImageIcon imageIcon = GuiUtils.bekommeSkaliertesImageIcon("images/fragezeichen.PNG", 39, 39);
 
 		JLabel lblBild = new JLabel();
 		lblBild.setIcon(imageIcon);
@@ -66,7 +78,7 @@ public class HilfeFenster {
 		panel.add(lblBild);
 
 		// Label-Bild 2
-		imageIcon = GuiUtils.getScaledImageIcon("images/hilfeIcon.PNG", 160, 160);
+		imageIcon = GuiUtils.bekommeSkaliertesImageIcon("images/hilfeIcon.PNG", 160, 160);
 
 		JLabel lblBild2 = new JLabel();
 		lblBild2.setIcon(imageIcon);
