+++ ProjektKlausurgenerator/src/de/arbeitsagentur/ProjektKlausurgenerator/model/Klausur.java
@@ -4,13 +4,12 @@ import java.util.List;
 
 public class Klausur {
 
-	private int anzahlPunkte;
 	private double punkteDouble;
 	private String klausurname;
 	private List<AbstractFrage> fragenList;
-	
-	public int getPunkte() {
-		return anzahlPunkte;
+
+	public double getPunkte() {
+		return punkteDouble;
 	}
 
 	public String getKlausurName() {
@@ -21,18 +20,10 @@ public class Klausur {
 		return fragenList;
 	}
 
-	public Klausur(int anzahlPunkte, String klausurname, List<AbstractFrage> fragenList){
-		this.anzahlPunkte = anzahlPunkte;
-		this.klausurname = klausurname;
-		this.fragenList = fragenList;
-	}
-	
-	public Klausur(Double punkteDouble, String klausurname, List<AbstractFrage> fragenList){
+	public Klausur(Double punkteDouble, String klausurname, List<AbstractFrage> fragenList) {
 		this.punkteDouble = punkteDouble;
 		this.klausurname = klausurname;
 		this.fragenList = fragenList;
 	}
-	
-	
-	
+
 }
